refactor: fix up file naming and comments
