+++ app-swift/Fushigi/FushigiApp.swift
@@ -11,9 +11,7 @@ import SwiftUI
 @main
 struct FushigiApp: App {
     var sharedModelContainer: ModelContainer = {
-        let schema = Schema([
-            Item.self,
-        ])
+        let schema = Schema([])
         let modelConfiguration = ModelConfiguration(schema: schema, isStoredInMemoryOnly: false)
 
         do {
