+++ app-swift/Fushigi/Stores/GrammarStore.swift
@@ -1,0 +1,7 @@
+//
+//  GrammarStore.swift
+//  Fushigi
+//
+//  Created by Tahoe Schrader on 2025/08/14.
+//
+
