+++ app-swift/Fushigi/Stores/JournalEntryStore.swift
@@ -1,0 +1,7 @@
+//
+//  JournalStore.swift
+//  Fushigi
+//
+//  Created by Tahoe Schrader on 2025/08/14.
+//
+
